feat: refactor chapter 6
