+++ content/chapters/chapter-6/index.tsx
@@ -27,9 +27,9 @@ export const metadata = {
     'put-it-together-1',
     'put-it-together-2',
     'put-it-together-3',
-    'put-it-together-4',
-    'put-it-together-5',
-    'put-it-together-6',
+    //'put-it-together-4',
+    //'put-it-together-5',
+    //'put-it-together-6',
   ],
   outros: ['outro-1'],
   challenges: ['in-out-1', 'put-it-together-1'],
