+++ content/lessons/chapter-6/put-it-together-2.tsx
@@ -1,14 +1,16 @@
 'use client'
-import { ScriptingChallenge, Table, Text, LessonInfo } from 'ui'
-import { useEffect, useState } from 'react'
-import { useTranslations } from 'hooks'
+
+import { ScriptingChallenge, LessonInfo, Title } from 'ui'
 import { EditorConfig } from 'types'
+import { useTranslations } from 'hooks'
+import { Text } from 'ui'
+import { useEffect, useState } from 'react'
 import { getLessonKey } from 'lib/progress'
 import { getData } from 'api/data'
 import { detectLanguage, Language } from 'lib/SavedCode'
 
 export const metadata = {
-  title: 'chapter_six.put_it_together_one.title',
+  title: 'chapter_six.put_it_together_four.title',
   key: 'CH6PUT2',
 }
 
@@ -16,12 +18,14 @@ export default function PutItTogether2({ lang }) {
   const t = useTranslations(lang)
   const [prevData, setPrevData] = useState<any>({ lesson: '', data: '' })
   const [isLoading, setIsLoading] = useState(true)
-
+  function countLines(text: string): number {
+    return text.split(/\r\n|\r|\n/).length
+  }
   const getPrevLessonData = async () => {
-    const data = await getData('CH6INO4')
+    const data = await getData('CH6PUT1')
     if (data) {
       setPrevData({
-        lesson_id: 'CH6INO4',
+        lesson_id: 'CH6PUT1',
         data: data?.code?.getDecoded(),
       })
     }
@@ -42,8 +46,7 @@ class Outpoint {
     assert(Number.isInteger(index));
     this.txid = txid;
     this.index = index;
-  }
-
+    }
   serialize() {
     const buf = Buffer.alloc(36);
     this.txid.copy(buf, 0);
@@ -51,7 +54,6 @@ class Outpoint {
     return buf;
   }
 }
-
 class Input {
   constructor() {
     this.outpoint = null;
@@ -60,7 +62,6 @@ class Input {
     this.value = 0;
     this.scriptcode = Buffer.alloc(0);
   }
-
   static from_output(txid, vout, value, scriptcode) {
     const self = new this();
     self.outpoint = new Outpoint(Buffer.from(txid.replace('0x', ''),  'hex').reverse(), vout);
@@ -68,7 +69,6 @@ class Input {
     self.scriptcode = Buffer.from(scriptcode.replace('0x', ''), 'hex');
     return self;
   }
-
   serialize() {
     const buf = Buffer.alloc(32 + 4 + 1 + 4);
     this.outpoint.serialize().copy(buf, 0);
@@ -80,14 +80,12 @@ class Input {
     return buf;
   }
 }
-
 class Output {
   constructor() {
     this.value = 0;
     this.witness_version = 0;
     this.witness_data = Buffer.alloc(0);
   }
-
   static from_options(addr, value) {
     assert(Number.isInteger(value));
     const self = new this();
@@ -97,7 +95,6 @@ class Output {
     self.value = value;
     return self;
   }
-
   serialize() {
     const buf = Buffer.alloc(11);
     buf.writeBigInt64LE(BigInt(this.value), 0);
@@ -107,6 +104,39 @@ class Output {
     return Buffer.concat([buf, this.witness_data]);
   }
 }
+const priv_dsljfohd = 0x93485bbe0f0b2810937fc90e8145b2352b233fbd3dd7167525401dd30738503en
+const compressed_pub_agfwuebb = Buffer.from('038cd0455a2719bf72dc1414ef8f1675cd09dfd24442cb32ae6e8c8bbf18aaf5af', 'hex');
+const txid_noiewnoa = '8a081631c920636ed71f9de5ca24cb9da316c2653f4dc87c9a1616451c53748e';
+const vout_bucbsncc = 1;
+const valueOne_aelfhasc = 650000000;
+const scriptcode_iabsvalb = '1976a914b234aee5ee74d7615c075b4fe81fd8ace54137f288ac';
+const input_bauoevbs = Input.from_output(txid_noiewnoa, vout_bucbsncc, valueOne_aelfhasc, scriptcode_iabsvalb);
+const addr_pqvejvea = 'bc1qgghq08syehkym52ueu9nl5x8gth23vr8hurv9dyfcmhaqk4lrlgs28epwj';
+const valueTwo_jhcermcr = 100000000;
+const output_uhmhvgcw = Output.from_options(addr_pqvejvea, valueTwo_jhcermcr);
+const tx_eagmcued = new Transaction();
+tx_eagmcued.inputs.push(input_bauoevbs);
+tx_eagmcued.outputs.push(output_uhmhvgcw);
+const [r, s] = tx_eagmcued.compute_input_signature(0, priv_dsljfohd);
+const EC = require('elliptic').ec;
+const ec = new EC('secp256k1');
+const private_key_faxwmufa = ec.keyFromPrivate(priv_dsljfohd.toString(16), 'hex');
+const hashed_message_bytes_ahuhfxmw = tx_eagmcued.digest(0);
+const signature_weiucmwa = {r: r.toString(16), s: s.toString(16)};
+console.log(private_key_faxwmufa.verify(hashed_message_bytes_ahuhfxmw, signature_weiucmwa) && 'true')
+console.log("KILL")`,
+    rangesToCollapse: [
+      {
+        start: countLines(prevData.data) + 55,
+        end: countLines(prevData.data) + 76,
+      },
+    ],
+    defaultFunction: {
+      name: 'privateKeyToPublicKey',
+      args: ['privateKey'],
+    },
+    defaultCode: `const secp256k1 = require('@savingsatoshi/secp256k1js');
+const {randomBytes} = require('crypto');
 
 class Witness {
   constructor() {
@@ -125,52 +155,81 @@ class Witness {
   }
 }
 
-const txid_fjpolkwe = '8a081631c920636ed71f9de5ca24cb9da316c2653f4dc87c9a1616451c53748e';
-const vout_eprwqlas = 1;
-const value1_vflkuyed = 650000000;
-const scriptcode_zlkwebnf = '1976a914b234aee5ee74d7615c075b4fe81fd8ace54137f288ac';
-const input_opjkfqsd = Input.from_output(txid_fjpolkwe, vout_eprwqlas, value1_vflkuyed, scriptcode_zlkwebnf);
-const addr_vjaudkfa = 'bc1qgghq08syehkym52ueu9nl5x8gth23vr8hurv9dyfcmhaqk4lrlgs28epwj';
-const value2_benadijl = 100000000;
-const output_tkuaojfp = Output.from_options(addr_vjaudkfa, value2_benadijl);
-const witness_qkadjyrf = new Witness();
-witness_qkadjyrf.push_item(Buffer.from('304402202e343143d5fcb0e3ece2ef11983d69dcaeb7407efe2ec7e3c830ab66927823c0022000ac4c1b3bcc857684e6bc2a36c07757695ef72b7bac70d2c877895798c4d1ba01', 'hex'));
-witness_qkadjyrf.push_item(Buffer.from('038cd0455a2719bf72dc1414ef8f1675cd09dfd24442cb32ae6e8c8bbf18aaf5af', 'hex'));
-const tx_dluitpjd = new Transaction();
-tx_dluitpjd.inputs.push(input_opjkfqsd);
-tx_dluitpjd.outputs.push(output_tkuaojfp);
-console.log(tx_dluitpjd.serialize().toString('hex')==='020000000001018e74531c4516169a7cc84d3f65c216a39dcb24cae59d1fd76e6320c93116088a0100000000ffffffff0100e1f50500000000220020422e079e04cdec4dd15ccf0b3fd0c742eea8b067bf06c2b489c6efd05abf1fd100000000'&&'true');
-`,
-    defaultFunction: {
-      name: 'put-it-together-2',
-      args: ['args'],
-    },
-    defaultCode: `class Transaction {
-  constructor() {
-    this.version=2;
-    this.flags = Buffer.from('0001', 'hex');
-    this.inputs = [];
-    this.outputs = [];
-    this.witnesses = [];
-    this.locktime = 0;
+${prevData.data.slice(0, -2)}
+  compute_input_signature(index, key) {
+    assert(typeof key === 'bigint');
+    assert(Number.isInteger(index));
+    // Helper function:
+    // Find modular multiplicative inverse using Extended Euclidean Algorithm
+    function invert(value, modulus = secp256k1.ORDER) {
+      let x0 = 0n;
+      let x1 = 1n;
+      let a = value;
+      let m = modulus;
+
+      while (a > 1n) {
+        const q = a / m;
+        let t = m;
+        m = a % m;
+        a = t;
+        t = x0;
+        x0 = x1 - q * x0;
+        x1 = t;
+      }
+
+      if (x1 < 0n)
+        x1 += modulus;
+        return x1;
+    }
+    // k = random integer in [1, n-1]
+    // R = G * k
+    // r = x(R) mod n
+    // s = (r * a + m) / k mod n
+    // Extra Bitcoin rule from BIP 146
+    // https://github.com/bitcoin/bips/blob/master/bip-0146.mediawiki#user-content-LOW_S
+    //   s = -s mod n, if s > n / 2
+    // return (r, s)
+    // YOUR CODE HERE
+    return [r, s];
   }
-  serialize() {
+  // Now sign the input you just returned!
+  sign_input(index, priv, pub, sighash=1) {
+    // Represent in DER format. The byte representations of r and s have
+    // length rounded up (255 bits becomes 32 bytes and 256 bits becomes 33 bytes).
+    // See BIP 66
+    // https://github.com/bitcoin/bips/blob/master/bip-0066.mediawiki
+    const der_int = (n) => {
+      const bit_length = n.toString(2).length;
+      const byte_length = parseInt((bit_length + 8) / 8);
+      const padded_hex = n.toString(16).padStart(byte_length * 2, '0');
+      return Buffer.from(padded_hex, 'hex');
+    };
+    const encode_der = (r, s) => {
+      const rb = der_int(r);
+      const sb = der_int(s);
+      let buf = Buffer.from([0x30, 4 + rb.length + sb.length, 2, rb.length]);
+      buf = Buffer.concat([buf, rb]);
+      buf = Buffer.concat([buf, Buffer.from([2, sb.length])]);
+      buf = Buffer.concat([buf, sb]);
+      return buf;
+    };
     // YOUR CODE HERE
   }
-}
-`,
+}`,
     validate: async (answer: string) => {
       if (answer) {
         if (answer === 'true') {
-          return [true, 'Nicely done']
+          return [true, '']
+        } else {
+          return [false, 'recheck your methods']
         }
-        return [false, 'Not a valid hex value']
+      } else {
+        return [false, "can't find a return in both of the methods"]
       }
-      return [false, 'Please return a value']
     },
     constraints: [
       {
-        range: [1, 1, 14, 1],
+        range: [1, 1, countLines(prevData.data.slice(0, -2)) + 76, 1],
         allowMultiline: true,
       },
     ],
@@ -244,6 +303,41 @@ class Output:
         r += self.witness_data
         return r
 
+priv_dsljfohd = 0x93485bbe0f0b2810937fc90e8145b2352b233fbd3dd7167525401dd30738503e
+compressed_pub_agfwuebb = bytes.fromhex("038cd0455a2719bf72dc1414ef8f1675cd09dfd24442cb32ae6e8c8bbf18aaf5af")
+txid_noiewnoa = "8a081631c920636ed71f9de5ca24cb9da316c2653f4dc87c9a1616451c53748e"
+valueOne_aelfhasc = 650000000
+scriptcode_iabsvalb = "1976a914b234aee5ee74d7615c075b4fe81fd8ace54137f288ac"
+vout_bucbsncc = 1
+input_bauoevbs = Input.from_output(txid_noiewnoa, vout_bucbsncc, valueOne_aelfhasc, scriptcode_iabsvalb)
+addr_pqvejvea = "bc1qgghq08syehkym52ueu9nl5x8gth23vr8hurv9dyfcmhaqk4lrlgs28epwj"
+valueTwo_jhcermcr = 100000000
+output_uhmhvgcw = Output.from_options(addr_pqvejvea, valueTwo_jhcermcr)
+tx_eagmcued = Transaction()
+tx_eagmcued.inputs.append(input_bauoevbs)
+tx_eagmcued.outputs.append(output_uhmhvgcw)
+(r, s) = tx_eagmcued.compute_input_signature(0, priv_dsljfohd)
+import ecdsa
+from ecdsa import VerifyingKey, SECP256k1
+sig_string_oiadhald = f'{r:064x}{s:064x}'
+sig_bytes_ayeqncas = bytes.fromhex(sig_string_oiadhald)
+hashed_message_bytes_ywienvsd = tx_eagmcued.digest(0)
+verifying_key_dojssdfo = VerifyingKey.from_string(compressed_pub_agfwuebb, curve=SECP256k1)
+print(verifying_key_dojssdfo.verify_digest(sig_bytes_ayeqncas, hashed_message_bytes_ywienvsd) and 'true')
+print("KILL")`,
+    rangesToCollapse: [
+      {
+        start: countLines(prevData.data) + 4,
+        end: countLines(prevData.data) + 16,
+      },
+    ],
+    defaultFunction: {
+      name: 'privatekey_to_publickey',
+      args: ['private_key'],
+    },
+    defaultCode: `from random import randrange
+from secp256k1py import secp256k1
+
 class Witness:
     def __init__(self):
         self.items = []
@@ -258,53 +352,46 @@ class Witness:
             r += pack("<B", len(item))
             r += item
         return r
+${prevData.data}
 
-txid_fjpolkwe = "8a081631c920636ed71f9de5ca24cb9da316c2653f4dc87c9a1616451c53748e"
-vout_eprwqlas = 1
-value_one_vflkuyed = 650000000
-scriptcode_zlkwebnf = "1976a914b234aee5ee74d7615c075b4fe81fd8ace54137f288ac"
-input_opjkfqsd = Input.from_output(txid_fjpolkwe, vout_eprwqlas, value_one_vflkuyed, scriptcode_zlkwebnf)
-addr_vjaudkfa = "bc1qgghq08syehkym52ueu9nl5x8gth23vr8hurv9dyfcmhaqk4lrlgs28epwj"
-value_two_benadijl = 100000000
-output_tkuaojfp = Output.from_options(addr_vjaudkfa, value_two_benadijl)
-witness_qkadjyrf = Witness()
-witness_qkadjyrf.push_item(bytes.fromhex("304402202e343143d5fcb0e3ece2ef11983d69dcaeb7407efe2ec7e3c830ab66927823c0022000ac4c1b3bcc857684e6bc2a36c07757695ef72b7bac70d2c877895798c4d1ba01"))
-witness_qkadjyrf.push_item(bytes.fromhex("038cd0455a2719bf72dc1414ef8f1675cd09dfd24442cb32ae6e8c8bbf18aaf5af"))
-tx_dluitpjd = Transaction()
-tx_dluitpjd.inputs.append(input_opjkfqsd)
-tx_dluitpjd.outputs.append(output_tkuaojfp)
-print(tx_dluitpjd.serialize().hex()=='020000000001018e74531c4516169a7cc84d3f65c216a39dcb24cae59d1fd76e6320c93116088a0100000000ffffffff0100e1f50500000000220020422e079e04cdec4dd15ccf0b3fd0c742eea8b067bf06c2b489c6efd05abf1fd100000000'and 'true')
-print("KILL")`,
-    defaultFunction: {
-      name: 'put-it-together-2',
-      args: ['args'],
-    },
-    defaultCode: `from struct import pack
+    def compute_input_signature(self, index: int, key: int):
+        # k = random integer in [1, n-1]
+        # R = G * k
+        # r = x(R) mod n
+        # s = (r * a + m) / k mod n
+        # Extra Bitcoin rule from BIP 146
+        # https://github.com/bitcoin/bips/blob/master/bip-0146.mediawiki#user-content-LOW_S
+        #   s = -s mod n, if s > n / 2
+        # return (r, s)
+        assert isinstance(key, int)
+        # YOUR CODE HERE
+        return (r, s)
 
-class Transaction:
-    def __init__(self):
-        self.version = 2
-        self.flags = bytes.fromhex("0001")
-        self.inputs = []
-        self.outputs = []
-        self.witnesses = []
-        self.locktime = 0
-
-    def serialize(self):
+    def sign_input(self, index, priv, pub, sighash=1):
+        def encode_der(r, s):
+            # Represent in DER format. Thebyte representations of r and s have length rounded up
+            # (255 bits becomes 32 bytes and 256 bits becomes 33 bytes).
+            # See BIP66
+            # https://github.com/bitcoin/bips/blob/master/bip-0066.mediawiki
+            rb = r.to_bytes((r.bit_length() + 8) // 8, 'big')
+            sb = s.to_bytes((s.bit_length() + 8) // 8, 'big')
+            return b'\\x30' + bytes([4 + len(rb) + len(sb), 2, len(rb)]) + rb + bytes([2, len(sb)]) + sb
         # YOUR CODE HERE
 `,
-    validate: async (answer: string) => {
+    validate: async (answer) => {
       if (answer) {
         if (answer === 'true') {
-          return [true, 'Nicely Done ']
+          return [true, '']
+        } else {
+          return [false, 'recheck your methods']
         }
-        return [false, 'Not a valid hex value']
+      } else {
+        return [false, "can't find a return in both of the methods"]
       }
-      return [false, 'Return a value']
     },
     constraints: [
       {
-        range: [1, 1, 14, 1],
+        range: [1, 1, countLines(prevData.data) + 15, 1],
         allowMultiline: true,
       },
     ],
@@ -320,6 +407,7 @@ class Transaction:
       python,
     },
   }
+
   return (
     !isLoading && (
       <ScriptingChallenge
@@ -327,79 +415,18 @@ class Transaction:
         config={config}
         saveData
         lessonKey={getLessonKey('chapter-6', 'put-it-together-2')}
-        successMessage={t('chapter_six.put_it_together_two.success')}
+        successMessage={t('chapter_six.put_it_together_four.success')}
       >
         <LessonInfo className="overflow-y-scroll  sm:max-h-[calc(100vh-70px)]">
-          <Text className="font-nunito text-2xl font-bold text-white">
-            {t('chapter_six.put_it_together_two.heading')}
+          <Title>{t('chapter_six.put_it_together_four.heading')}</Title>
+          <Text className="mt-4 font-nunito text-xl text-white">
+            {t('chapter_six.put_it_together_four.paragraph_one')}
           </Text>
-          <Text className="mt-2 font-nunito text-xl text-white">
-            {t('chapter_six.put_it_together_two.paragraph_one')}
+          <Text className="mt-4 font-nunito text-xl text-white">
+            {t('chapter_six.put_it_together_four.paragraph_two')}
           </Text>
-          <Text className="mt-4 text-lg md:text-xl">
-            {t('chapter_six.put_it_together_two.paragraph_two')}
-          </Text>
-          <Table
-            headings={[
-              t('chapter_six.put_it_together_two.headings.item_one'),
-              t('chapter_six.put_it_together_two.headings.item_two'),
-              t('chapter_six.put_it_together_two.headings.item_three'),
-              t('chapter_six.put_it_together_two.headings.item_four'),
-            ]}
-            rows={[
-              [
-                t('chapter_six.put_it_together_two.table.row_one.item_one'),
-                t('chapter_six.put_it_together_two.table.row_one.item_two'),
-                t('chapter_six.put_it_together_two.table.row_one.item_three'),
-                t('chapter_six.put_it_together_two.table.row_one.item_four'),
-              ],
-              [
-                t('chapter_six.put_it_together_two.table.row_two.item_one'),
-                t('chapter_six.put_it_together_two.table.row_two.item_two'),
-                t('chapter_six.put_it_together_two.table.row_two.item_three'),
-                t('chapter_six.put_it_together_two.table.row_two.item_four'),
-              ],
-              [
-                t('chapter_six.put_it_together_two.table.row_three.item_one'),
-                t('chapter_six.put_it_together_two.table.row_three.item_two'),
-                t('chapter_six.put_it_together_two.table.row_three.item_three'),
-                t('chapter_six.put_it_together_two.table.row_three.item_four'),
-              ],
-              [
-                t('chapter_six.put_it_together_two.table.row_four.item_one'),
-                t('chapter_six.put_it_together_two.table.row_four.item_two'),
-                t('chapter_six.put_it_together_two.table.row_four.item_three'),
-                t('chapter_six.put_it_together_two.table.row_four.item_four'),
-              ],
-              [
-                t('chapter_six.put_it_together_two.table.row_five.item_one'),
-                t('chapter_six.put_it_together_two.table.row_five.item_two'),
-                t('chapter_six.put_it_together_two.table.row_five.item_three'),
-                t('chapter_six.put_it_together_two.table.row_five.item_four'),
-              ],
-              [
-                t('chapter_six.put_it_together_two.table.row_six.item_one'),
-                t('chapter_six.put_it_together_two.table.row_six.item_two'),
-                t('chapter_six.put_it_together_two.table.row_six.item_three'),
-                t('chapter_six.put_it_together_two.table.row_six.item_four'),
-              ],
-              [
-                t('chapter_six.put_it_together_two.table.row_seven.item_one'),
-                t('chapter_six.put_it_together_two.table.row_seven.item_two'),
-                t('chapter_six.put_it_together_two.table.row_seven.item_three'),
-                t('chapter_six.put_it_together_two.table.row_seven.item_four'),
-              ],
-              [
-                t('chapter_six.put_it_together_two.table.row_eight.item_one'),
-                t('chapter_six.put_it_together_two.table.row_eight.item_two'),
-                t('chapter_six.put_it_together_two.table.row_eight.item_three'),
-                t('chapter_six.put_it_together_two.table.row_eight.item_four'),
-              ],
-            ]}
-          />
-
-          <Text className="mt-2 font-nunito text-xl text-white">
-            {t('chapter_six.put_it_together_two.paragraph_three')}
+          <Text className="mt-4 font-nunito text-xl text-white">
+            {t('chapter_six.put_it_together_four.paragraph_three')}
           </Text>
         </LessonInfo>
       </ScriptingChallenge>
